requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
pdf2image
faiss-cpu
langchain_google_genai
Pillow